+++ src/headless/store/enums/index.ts
@@ -1,3 +1,2 @@
 export * from './sort-enums';
-export * from './product-status-enums';
 export * from './social-platform-enums';
